TP_2/requirements.txt

aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.3.0
multiprocessing-logging==0.3.1
selenium==4.24.0
Pillow==10.3.0
requests==2.32.3
orjson==3.10.7
psutil
pytest
